moved file_comparison to validate and minoor updates
