+++ src/micromet/__init__.py
@@ -27,7 +27,6 @@ from .report import gap_summary
 from .format import transformers
 from .format import merge
 from .qaqc import data_cleaning
-from .qaqc import file_comparison
 from .format.transformers import columns, timestamps, validation, corrections, cleanup, interval_updates, MISSING_VALUE
 
 __version__ = "0.3.1"
